automation/pyproject.toml
[project]
name = "keycloak-automation"
version = "0.1.0"
description = "User lifecycle automation (onboarding & offboarding) for Keycloak and GitLab"
requires-python = ">=3.11"
dependencies = [
    "python-keycloak>=3.7.0",
    "python-gitlab>=4.4.0",
    "click>=8.1.7",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "apscheduler>=3.10.4",
]

[dependency-groups]
dev = []
